refactor: split sparse and dense parameters
