+++ vowpalwabbit/core/CMakeLists.txt
@@ -4,6 +4,7 @@ set(vw_core_headers
   include/vw/core/active_multiclass_prediction.h
   include/vw/core/api_status.h
   include/vw/core/array_parameters_dense.h
+  include/vw/core/array_parameters_sparse.h
   include/vw/core/array_parameters.h
   include/vw/core/beam.h
   include/vw/core/best_constant.h
@@ -21,6 +22,7 @@ set(vw_core_headers
   include/vw/core/continuous_actions_reduction_features.h
   include/vw/core/correctedMath.h
   include/vw/core/cost_sensitive.h
+  include/vw/core/cressieread_estimator.h
   include/vw/core/crossplat_compat.h
   include/vw/core/debug_log.h
   include/vw/core/debug_print.h
@@ -42,6 +44,7 @@ set(vw_core_headers
   include/vw/core/guard.h
   include/vw/core/hashstring.h
   include/vw/core/interactions_predict.h
+  include/vw/core/interactions.h
   include/vw/core/io_buf.h
   include/vw/core/json_utils.h
   include/vw/core/kskip_ngram_transformer.h
@@ -73,7 +76,6 @@ set(vw_core_headers
   include/vw/core/print_utils.h
   include/vw/core/prob_dist_cont.h
   include/vw/core/queue.h
-  include/vw/core/thread_pool.h
   include/vw/core/rand_state.h
   include/vw/core/rand48.h
   include/vw/core/reduction_features.h
@@ -131,7 +133,6 @@ set(vw_core_headers
   include/vw/core/reductions/get_pmf.h
   include/vw/core/reductions/interact.h
   include/vw/core/reductions/interaction_ground.h
-  include/vw/core/interactions.h
   include/vw/core/reductions/kernel_svm.h
   include/vw/core/reductions/lda_core.h
   include/vw/core/reductions/log_multi.h
@@ -169,7 +170,6 @@ set(vw_core_headers
   include/vw/core/reductions/svrg.h
   include/vw/core/reductions/topk.h
   include/vw/core/scope_exit.h
-  include/vw/core/cressieread_estimator.h
   include/vw/core/shared_data.h
   include/vw/core/simple_label_parser.h
   include/vw/core/simple_label.h
@@ -177,6 +177,7 @@ set(vw_core_headers
   include/vw/core/stable_unique.h
   include/vw/core/tag_utils.h
   include/vw/core/text_utils.h
+  include/vw/core/thread_pool.h
   include/vw/core/unique_sort.h
   include/vw/core/v_array.h
   include/vw/core/version.h
@@ -191,6 +192,8 @@ set(vw_core_sources
   src/accumulate.cc
   src/action_score.cc
   src/api_status.cc
+  src/array_parameters_dense.cc
+  src/array_parameters_sparse.cc
   src/best_constant.cc
   src/cache.cc
   src/cb_continuous_label.cc
@@ -200,6 +203,7 @@ set(vw_core_sources
   src/ccb_reduction_features.cc
   src/confidence_sequence.cc
   src/cost_sensitive.cc
+  src/cressieread_estimator.cc
   src/crossplat_compat.cc
   src/debug_print.cc
   src/decision_scores.cc
@@ -210,9 +214,11 @@ set(vw_core_sources
   src/gen_cs_example.cc
   src/global_data.cc
   src/hashstring.cc
+  src/interactions.cc
   src/io_buf.cc
   src/kskip_ngram_transformer.cc
   src/label_dictionary.cc
+  src/label_parser.cc
   src/label_parser.cc
   src/label_type.cc
   src/learner.cc
@@ -238,10 +244,6 @@ set(vw_core_sources
   src/reductions/active.cc
   src/reductions/audit_regressor.cc
   src/reductions/autolink.cc
-  src/reductions/details/automl/automl_impl.cc
-  src/reductions/details/automl/automl_oracle.cc
-  src/reductions/details/automl/automl_iomodel.cc
-  src/reductions/details/automl/automl_util.cc
   src/reductions/automl.cc
   src/reductions/baseline_challenger_cb.cc
   src/reductions/baseline.cc
@@ -278,6 +280,10 @@ set(vw_core_sources
   src/reductions/cs_active.cc
   src/reductions/csoaa_ldf.cc
   src/reductions/csoaa.cc
+  src/reductions/details/automl/automl_impl.cc
+  src/reductions/details/automl/automl_iomodel.cc
+  src/reductions/details/automl/automl_oracle.cc
+  src/reductions/details/automl/automl_util.cc
   src/reductions/ect.cc
   src/reductions/epsilon_decay.cc
   src/reductions/explore_eval.cc
@@ -289,7 +295,6 @@ set(vw_core_sources
   src/reductions/get_pmf.cc
   src/reductions/interact.cc
   src/reductions/interaction_ground.cc
-  src/interactions.cc
   src/reductions/kernel_svm.cc
   src/reductions/lda_core.cc
   src/reductions/log_multi.cc
@@ -326,7 +331,6 @@ set(vw_core_sources
   src/reductions/stagewise_poly.cc
   src/reductions/svrg.cc
   src/reductions/topk.cc
-  src/cressieread_estimator.cc
   src/shared_data.cc
   src/simple_label_parser.cc
   src/simple_label.cc
