+++ vowpalwabbit/core/include/vw/core/global_data.h
@@ -36,8 +36,6 @@
 #  include <thread>
 #endif
 
-using weight = float;
-
 using feature_dict = std::unordered_map<std::string, std::unique_ptr<features>>;
 using reduction_setup_fn = VW::LEARNER::base_learner* (*)(VW::setup_base_i&);
